[mypyc] Raise NameError on undefined names (#19395)
Fixes https://github.com/mypyc/mypyc/issues/879

Changed the type of exception that is raised on undefined names from
`RuntimeError` to `NameError`.

Moved the runtime error that refers to unexpected execution of
unreachable code to `shortcircuit_expr` to keep existing behavior and
generate the exception in blocks that are unreachable due to
short-circuiting.

In the future this could probably be optimized to not generate any code
at all and just constant-fold the boolean expression.