[mypyc] Fail run test if default driver does not find test cases (#19420)
If a `run-` test doesn't define a custom `driver.py`, the default one
looks for functions prefixed `test_` and calls them. If the prefix is
missing or misspelled, or there is some other issue that causes the test
to have no test cases, the test succeeds without running the cases.

To try to prevent this, the default driver will now fail if it doesn't
find any test cases. Existing tests are changed to conform to this
requirement in this PR.

For most tests that meant simply moving the statements from top level to
a function, but some were revealed to have been broken and weren't
actually run, for example because `[typing fixtures/typing-full.pyi]`
was at the top of the test instead of at the bottom, which made the test
setup code ignore all of the test case.