[mypyc] Report error when registering a nested function (#19450)
Fixes https://github.com/mypyc/mypyc/issues/1118

Added a check for nested `@singledispatch` functions and nested
functions registered to `@singledispatch` functions to report an error
in mypyc. Currently either of those cases causes mypyc to crash because
of the different handling of nested and top-level functions.

Changed to abort the compilation early when these errors are found so
that in the transform code we can assume that the singledispatch
functions are valid. This means that mypyc might not report as many
errors until it quits as before, so I have split the error output test
in `commandline.test` into two.