[mypyc] Only generate an export table if using separate compilation (#19521)
When not using separate compilation, the export table is not used. Also,
there's actually no simple and safe way to use it without separate
compilation, since we'd have to first ensure that the structure of the
export table is compatible, as otherwise the order of fields or the
types of the fields could be incompatible and cause segfaults and other
fun stuff.