Try simple-minded call expression cache (#19505)
This gives a modest 1% improvement on self-check (compiled), but it
gives almost 40% on `mypy -c "import colour"`. Some comments:
* I only cache `CallExpr`, `ListExpr`, and `TupleExpr`, this is not very
principled, I found this as a best balance between rare cases like
`colour`, and more common cases like self-check.
* Caching is fragile within lambdas, so I simply disable it, it rarely
matters anyway.
* I cache both messages and the type map, surprisingly the latter only
affects couple test cases, but I still do this generally for peace of
mind.
* It looks like there are only three things that require cache
invalidation: binder, partial types, and deferrals.

In general, this is a bit scary (as this a major change), but also perf
improvements for slow libraries are very tempting.