Use cache for OpExpr (#19523)
Fixes https://github.com/python/mypy/issues/14978

This is irrelevant for self-check (and for most code likely), but it is
important for numerical code, where it avoids exponential slowdown in
formulas involving arrays etc (see e.g. the original issue).

Unless there will be fallout in `mypy_primer` and/or there are
objections, I am going to go ahead and merge it.