[mypyc] Enable free threading when compiling multiple modules (#19541)
Previously we used multi-phase initialization to enable free threading
on Python builds that support it, but only if a single module was
compiled in a group.

Implements it also for multiple modules in a group. Add support for
multi-phase initialization in module shims and the shared library. It's
still only used on free-threaded builds, and we fall back to the old
approach on other Python versions/builds.

This enables compiling mypy and mypyc on free-threaded Python builds. At
least almost all mypy and mypyc tests now pass when compiled and on
3.14.0b4 with free threading (only tested on macOS so far).