Fix async mypyc tests on Windows
The async tests previously used many invocations of `asyncio.run`. This
could cause the test processes to die for strange event loop reasons.
This commit converts most test cases to be run in a single event loop
managed by the default driver, which is now async aware.

Not all tests could be converted, e.g. the test that runs an async
function in a sync context. However, the test suite does succeed with
these changes.