[mypyc] feat: unwrap NewType types to their base types for optimized code paths (#19497)
This PR adds special case logic for unwrapping NewType types to their
actual type. This logic is currently working. a `NewType("name", str)`
now generates the same code as a `str`.

I wasn't entirely sure of the best way to test this, so I just tweaked
the str tests to use a union of str and newtype str and validated that
the IR still uses `str` and not `object`

Almost all of my tests are running fine, but I get a strange mypy error
in the str.count tests saying that `str` objects do not have a .count
method? Of course a str object has a .count method. Do you think this
might related to the typeshed stuff again?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>