[mypyc] Add prefix to attributes of generator classes (#19535)
Fixes https://github.com/mypyc/mypyc/issues/1120

`async` functions were recently changed to be represented as generator
classes in the IR. Their parameters are represented as attributes in
those classes.

There are several methods added to every generator class by the
compiler, one of those is called `send`. If there is also a parameter
called `send`, mypyc assumes that the method is a property getter/setter
and inserts method calls in the generated code for `GetAttr` and
`SetAttr` nodes in the IR.

This is incorrect and led to the compilation error in the linked issue
because the `send` method of generator classes takes one more argument
than a property getter would.

The name clash is fixed by adding a prefix
`__mypyc_generator_attribute__` to attribute names derived from
parameters, which should ensure that argument references are not
converted into method calls.