Another two micro-optimizations (#19633)
Here are two things:
* Make `FormalArgument` a native class. We create huge amount of these
(as callable subtyping is one of the most common subtype checks), and
named tuples creation is significantly slower than native classes.
* Do not call `re.match()` in a code path of `format_type()`. This is
relatively slow (as it is a `py_call()`) and it is called in almost
every error message. This creates problems for code with many
third-party dependencies where these errors are ignored anyway.

FWIW in total these give ~0.5% together (I didn't measure individually,
but I guess the most benefit for self-check is from the first one).

---------

Co-authored-by: Ali Hamdan <ali.hamdan.dev@gmail.com>