Special-case enum method calls (#19634)
Improves https://github.com/mypyc/mypyc/issues/1121, this gives a bit
above 1% on mypy self-check.

This only adds support for regular and overloaded methods without
decorators (class/static methods and properties stay slow). When working
on this I considered (and actually tried) four options:
* Make enums extension classes, then many methods will use fast calls
~automatically (we will just need to set a final flag). This just didn't
work, in the sense no segfaults, but it looks like we don't call
`__prepare__()`, or don't call it at the right moment. Or maybe I just
didn't try hard enough. In general, for some reason this feels risky.
* Use existing `CPyDef`s for (non-extension) enum methods, but since
they have an extra argument, `__mypyc_self__`, we can supply `NULL`
there, since we know it is unused. This is actually easy and it works,
but IMO it is ultra-ugly, so I decided to not do it.
* Write a separate `CPyDef` without `__mypy_self__`, use it for direct
calls, and make existing callable classes `CPyDef`s one-line functions
that simply call the first one. This is possible, but quite complicated,
and I am not sure it is easy to generalize (e.g. on classmethods).
* Finally, the way I do this is to simply generate a second method, that
is almost a copy of the original one. This involves a bit of code
duplication (in C), but the benefit is that it is conceptually simple,
and easily extendable. We can cover more special cases on as-needed
basis.