[mypyc] feat: exact_dict_set_item_op (#19657)
This PR implements a `exact_dict_set_item_op` custom_op in preparation
for the addition of exact_dict_rprimitive. We don't actually need to
implement exact_dict_rprimitive before implementing this op as long as
we accept that it will not be used for CallExpr specialization. This is
okay for us now, as it is already addressed in the larger PR.