Fix spurious possibly-undefined errors in for-else with break
When a for loop contains branches with `break` and an `else` block,
variables declared inside those branches were incorrectly discarded
from further analysis, leading Mypy to incorrectly report a variable
as undefined after the loop or as used before declaration.

With this fix, when a for loop's `else` block is considered, variables
declared in every branch of the `for` loop body that called `break` are
now considered as defined within the body of the loop.

Fixes #14209
Fixes #19690