stubtest: get better signatures for `__init__` of C classes (#18259)
When an `__init__` method has the generic C-class signature, check the
underlying class for a better signature.

I was looking at `asyncio.futures.Future.__init__` and wanted to take
advantage of the better `__text_signature__` on
`asyncio.futures.Future`.

The upside is that this clears several allowlist entries and surfaced
several other cases where typeshed is currently incorrect, with no false
positives.