stubtest: handle overloads with mixed pos-only params (#18287)
Fixes #17023 

Stubtest should only mangle positional-only parameter names if they're
positional-only in all branches of the overload. The signatures get
really ugly and wrong otherwise.

I'm not sure if I did the new `test_overload_signature` in the best way.
I couldn't figure out a way to get covert a string into a
`nodes.OverloadedFuncDef` object with any of the techniques in existing
tests in `teststubtest.py`. Maybe the new test case is sufficient, but I
wanted to test the signature generation directly.