Remove unnecessary error message for match class patterns (#19708)
Remove `Cannot determine type of ...` error for class patterns if the
class to match cannot be resolved. In these cases a `Name ... is not
defined` error is already emitted. The captured variable should simply
be inferred as `Any`.

Previously, this was especially an issue for class matches to a class
from an untyped library together with a MemberExpr. An example from
pylint / astroid which shouldn't emit any errors:
```py
from typing import Any
from astroid import nodes

def func(var: Any) -> None:
    match var:
        case nodes.Assign(targets=t):
            reveal_type(t)  # Any
```