[mypyc] feat: PyObject_CallObject op for fn(*args) fastpath (#19631)
This PR adds a new custom_op for PyObject_CallObject which is more
efficient than PyObject_Call in cases where there are no kwargs.
posarg-only use cases are already optimized but this is helpful for
patterns such as `fn(*args)` or `fn(a1, a2, *args)`

This PR extends #19623 and #19629 , as this change will not be helpful
until those PRs are merged.