Use 1 byte per type/symbol tag (#19735)
This is a small incremental improvement for fixed format cache. I am
adding a dedicated write/read functions for tags (i.e. integers in 0-255
range). I propose to exclusively use these functions for type tags
(hence the name), and still use regular `write_int()`/`read_int()` for
integers that are "accidentally small" (like argument kinds etc). In a
separate PR I will change regular `int` format to be more progressive
(e.g. only use 1 byte if an integer happens to be small). I also change
the terminology from "marker" to "tag", as this is a more common name
for this concept.

Note we can probably use `mypy_extensions.u8` for type tags. If there is
a desire for this, I can switch to it (either in this or a separate PR).