[mypyc] feat: `__mypyc_empty_tuple__` constant (#19654)
I realized that any time a user has a kwarg-only call expression like
`fn(abc=123, ...)` in their compiled code, and `func` is not a native
function, a new empty tuple is created every time

This is not really necessary, we can just hold the same empty tuple in
memory as a constant and pass it around. It's immutable, and that's
already what we're already doing, since `tuple() is tuple()` but our
current method involves more steps.

This should slightly improve the speed of kwarg-only python func
calling.