Do not use outer context for `or` expr inference if the LHS has Any (#19748)
Fixes #19492.

In #19695 I tried to add another set of heuristics to `any_constraints`,
but can't get that working: trying to join/meet all similar constraints
together breaks inference in other cases, and only doing that for Any
would be somewhat non-trivial. This PR reverts behaviour introduced in
#19492 when LHS of the expression contains `Any`. Outer context is still
included in all other cases, and this seems to strike a good balance.