Use more compact cache representation for int and str (#19750)
After looking more at some real data I found that:
* More than 99.9% of all `int`s are between -10 and 117. Values are a
bit arbitrary TBH, the idea is that we should include small negative
values (for `TypeVarId`s) and still be able to fit them in 1 byte.
* More than 99.9% of strings are shorter that 128 bytes (again the idea
is to fit the length into a single byte)

Note there are very few integers that would fit in two bytes currently.
This is because we only store line for type alias nodes, and type
aliases are usually defined at the top of a module. We can add special
case for two bytes later when needed.

We could probably save another byte for long strings and medium
integers, but I don't want to have anything fancy that would only affect
less than 0.1% cases.

Finally you may notice I add a small correctness change I noticed
accidentally when working on this, it is not really related, but it is
so minor that it doesn't deserve a separate PR.