[mypyc] Speed up equality with optional str/bytes types (#19758)
Specialize most equality (`==` and `!=`) operations when one of the
operands is `str | None` or `bytes | None`. First check if the value is
`None`, and based on that branch into fast path operations. Previously
we used a generic C API primitive for such comparisons, which was quite
slow.

This could be generalized to other optional types, but let's start with
`str | None` since it's a very common type and it's often used in
equality tests. `bytes | None` is also covered, since it's very similar
to the `str` case.

Also add support for unchecked `Cast` operations in the IR. These don't
perform a runtime type check -- they can be used to narrow the static
type when it can be known statically that the cast is always safe.