Add await to empty context hack (#19777)
Fixes #19716. It is a follow-up to #19767 and was missed there due to
malformed test stubs (the same testcase fails when run by full mypy
against typeshed, I did not notice missing AwaitExpr because of the
passing test...). I synced typevar variance with typeshed definitions
and added AwaitExpr to the list of context-dependent exprs. Cc
@ilevkivskyi