Inverse interface freshness logic (#19809)
Fixes https://github.com/python/mypy/issues/9554

This is another case where I am surprised id didn't work like this in
the first place. Right now the freshness info originates from the
dependency itself (like trust me, I am fresh, whatever it means). IMO
this doesn't make much sense, instead a dependent should verify whether
all dependencies are the same it last seen them. On the surface the idea
is simple, but there are couple tricky parts:
* This requires splitting `write_cache()` in two phases: first write all
data files in an SCC (or at least serialize them), the write all meta
files. I didn't find any elegant way to do the split, but it is probably
fine, as we already have this untyped meta JSON in few places.
* I am adding plugin data (used by mypyc separate compilation currently)
as part of interface hash. It is not documented whether it should be
this way or not, but I would say it should, and this is essentially how
mypyc expects it (group name will appear in `#include
<__native_group_name.h>`, so it _is_ a part of the interface). It used
to work ~accidentally because we check plugin data in
`find_cache_meta()` that is called before setting `interface_hash`, not
in `validate_meta()`.