Rename optional arguments to utility arguments.
The rationale behind this being that all of our arguments are optional. But these are options that deal with the program as a utility (eg, version number) so that's a more fitting name.

The old anchor is recreated to aid incoming old links.

I have manually verified that the --help looks right.