[mypyc] Transform object.__new__ inside __new__ (#19866)
#19739 introduced support for compiling `__new__` methods in native
classes. Native classes differ internally from regular python types
which results in `TypeError`s being raised when `object.__new__(cls)` is
called when `cls` is a native class.

To avoid making this call, `super().__new__(cls)` is transformed into a
call to an internal setup function. I forgot to replicate this for calls
to equivalent `object.__new__(cls)` calls so this PR fixes that. This
introduced a regression because before my changes, `__new__` methods
with `object.__new__(cls)` were effectively ignored at runtime, and
after my changes they started raising `TypeError`s.

Note that these calls are left as-is outside of `__new__` methods so
it's still possible to trigger the `TypeError` but that is not a
regression as this was the case before. For example this code:
```
class Test:
    pass

t = object.__new__(Test)
```
results in `TypeError: object.__new__(Test) is not safe, use
Test.__new__()`. This differs from interpreted python but the error
message is actually correct in that using `Test.__new__(Test)` instead
works.