[mypyc] Generate __getattr__ wrapper (#19909)
Fixes https://github.com/mypyc/mypyc/issues/887

Generate a wrapper function for `__getattr__` in classes where it's
defined and put it into the `tp_getattro` slot.

At runtime, this wrapper function is called for every attribute access,
so to match behavior of interpreted python it needs to first check if
the given name is present in the type dictionary and only call the
user-defined `__getattr__` when it's not present.

Checking the type dictionary is implemented using a cpython function
`_PyObject_GenericGetAttrWithDict` which is also used in the default
attribute access handler in
[cpython](https://github.com/python/cpython/blob/dd45179fa0f5ad2fd169cdd35065df2c3bce85bc/Objects/typeobject.c#L10676).

In compiled code, the wrapper will only be called when the attribute
name cannot be statically resolved. When it can be resolved, the
attribute will be accessed directly in the underlying C struct, or the
generated function will be directly called in case of resolving method
names. No change from existing behavior.

When the name cannot be statically resolved, mypyc generates calls to
`PyObject_GetAttr` which internally calls `tp_getattro`.

In interpreted code that uses compiled classes, attribute access will
always result in calls to `PyObject_GetAttr` so there's always a dict
look-up in the wrapper to find the attribute. But that's the case
already, the dict look-up happens in the default attribute access
handler in cpython. So the wrapper should not bring any negative
performance impact.